# Fix bug in datagram message due to shared buffer